requirements.txt
numpy
scipy
pandas
librosa==0.10.2.post1
soundfile
tqdm
pyyaml
pyloudnorm
torch==2.3.1
torchaudio==2.3.1
torchcrepe
torchvision==0.18.1

laion-clap
faiss-cpu
sentence-transformers

scikit-learn